+++ lib/Db/Artist.php
@@ -9,7 +9,7 @@
  * @author Morris Jobke <hey@morrisjobke.de>
  * @author Pauli Järvinen <pauli.jarvinen@gmail.com>
  * @copyright Morris Jobke 2013, 2014
- * @copyright Pauli Järvinen 2017 - 2021
+ * @copyright Pauli Järvinen 2017 - 2023
  */
 
 namespace OCA\Music\Db;
@@ -28,8 +28,8 @@ use OCP\IURLGenerator;
  * @method void setHash(string $hash)
  * @method ?string getStarred()
  * @method void setStarred(?string $timestamp)
- * @method ?string getLastfmUrl()
- * @method void setLastfmUrl(?string $lastfmUrl)
+ * @method ?int getRating()
+ * @method setRating(?int $rating)
  */
 class Artist extends Entity {
 	public $name;
@@ -37,14 +37,24 @@ class Artist extends Entity {
 	public $mbid;
 	public $hash;
 	public $starred;
+	public $rating;
 
 	// not part of the standard content, injected separately when needed
-	public $lastfmUrl;
+	private $lastfmUrl;
 
 	public function __construct() {
 		$this->addType('coverFileId', 'int');
+		$this->addType('rating', 'int');
 	}
 
+	public function getLastfmUrl() : ?string {
+		return $this->lastfmUrl;
+	}
+
+	public function setLastfmUrl(?string $lastfmUrl) : void {
+		$this->lastfmUrl = $lastfmUrl;
+	}
+   
 	public function getUri(IURLGenerator $urlGenerator) {
 		return $urlGenerator->linkToRoute(
 			'music.shivaApi.artist',